# apps/api/openapi/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/notes": {
      "post": {
        "operationId": "NotebookController_createNote",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNoteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TechnicalNoteDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "interviewos_session": []
          }
        ],
        "summary": "Create a technical note draft",
        "tags": [
          "notes"
        ]
      },
      "get": {
        "operationId": "NotebookController_findNotes",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TechnicalNoteDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "interviewos_session": []
          }
        ],
        "summary": "List the current user technical notes",
        "tags": [
          "notes"
        ]
      }
    },
    "/notes/{noteId}": {
      "get": {
        "operationId": "NotebookController_findNoteById",
        "parameters": [
          {
            "name": "noteId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TechnicalNoteDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "interviewos_session": []
          }
        ],
        "summary": "Get a technical note by id",
        "tags": [
          "notes"
        ]
      },
      "patch": {
        "operationId": "NotebookController_updateNote",
        "parameters": [
          {
            "name": "noteId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNoteDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TechnicalNoteDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "interviewos_session": []
          }
        ],
        "summary": "Update a technical note",
        "tags": [
          "notes"
        ]
      },
      "delete": {
        "operationId": "NotebookController_deleteNote",
        "parameters": [
          {
            "name": "noteId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TechnicalNoteDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "interviewos_session": []
          }
        ],
        "summary": "Archive a technical note",
        "tags": [
          "notes"
        ]
      }
    },
    "/notes/{noteId}/generate": {
      "post": {
        "operationId": "NotebookController_generateTechnicalNote",
        "parameters": [
          {
            "name": "noteId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TechnicalNoteDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "interviewos_session": []
          }
        ],
        "summary": "Generate structured content for a technical note",
        "tags": [
          "notes"
        ]
      }
    },
    "/notes/{noteId}/questions": {
      "post": {
        "operationId": "NotebookController_generateQuestions",
        "parameters": [
          {
            "name": "noteId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateQuestionsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TechnicalNoteDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "interviewos_session": []
          }
        ],
        "summary": "Generate interview questions from a technical note",
        "tags": [
          "notes"
        ]
      }
    },
...
    "/review": {
      "get": {
        "operationId": "ReviewController_getReviewQueue",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewQueueResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "interviewos_session": []
          }
        ],
        "summary": "Get the prioritized review queue",
        "tags": [
          "review"
        ]
      }
    },
    "/review/{reviewItemId}/rate": {
      "post": {
        "operationId": "ReviewController_rateReviewItem",
        "parameters": [
          {
            "name": "reviewItemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewRatingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewAttemptDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "interviewos_session": []
          }
        ],
        "summary": "Rate a review item and schedule the next review",
        "tags": [
          "review"
        ]
      }
    },
    "/learning-path": {
      "get": {
        "operationId": "ReviewController_getLearningPath",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LearningPathItemResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "interviewos_session": []
          }
        ],
        "summary": "Build the current learning path from review signals",
        "tags": [
          "review"
        ]
      }
    },
    "/learning-path/{itemId}/action": {
      "post": {
        "operationId": "ReviewController_updateLearningPathItem",
        "parameters": [
          {
            "name": "itemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LearningPathActionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningPathItemResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "interviewos_session": []
          }
        ],
        "summary": "Update a learning path item status",
        "tags": [
          "review"
        ]
      }
    },
    "/weak-concepts": {
      "get": {
        "operationId": "ReviewController_listWeakConcepts",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WeakConceptResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "interviewos_session": []
          }
        ],
        "summary": "List weak concepts detected from practice sessions",
        "tags": [
          "review"
        ]
      }
    },
    "/weak-concepts/{weakConceptId}/status": {
      "patch": {
        "operationId": "ReviewController_updateWeakConceptStatus",
        "parameters": [
          {
            "name": "weakConceptId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WeakConceptStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeakConceptResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "interviewos_session": []
          }
        ],
        "summary": "Update the status of a weak concept",
        "tags": [
          "review"
        ]
      }
    },
...
  },
  "info": {
    "title": "InterviewOS API",
    "description": "REST API for InterviewOS local development and integration testing.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "interviewos_session"
      }
    },
    "schemas": {
...
      "TechnicalNoteContentDto": {
        "type": "object",
        "properties": {
          "purpose": {
            "type": "string"
          },
          "quickReference": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "coreConcepts": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "mentalModel": {
            "type": "string"
          },
          "productionUsage": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "practicalExamples": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "commonPitfalls": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "debuggingChecklist": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "productionChecklist": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "seniorInterviewSignals": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "purpose",
          "quickReference",
          "coreConcepts",
          "mentalModel",
          "productionUsage",
          "practicalExamples",
          "commonPitfalls",
          "debuggingChecklist",
          "productionChecklist",
          "seniorInterviewSignals"
        ]
      },
      "TechnicalNoteContentSectionDto": {
        "type": "object",
        "properties": {
          "heading": {
            "type": "string"
          },
          "content": {
            "type": "string"
          }
        },
        "required": [
          "heading",
          "content"
        ]
      },
      "NoteGeneratedQuestionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "noteId": {
            "type": "string"
          },
          "question": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "expectedAnswer": {
            "type": "string",
            "nullable": true
          },
          "difficulty": {
            "type": "string",
            "enum": [
              "EASY",
              "MEDIUM",
              "HARD",
              "EXPERT"
            ]
          },
          "expectedConcepts": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "sourceSection": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "noteId",
          "question",
          "category",
          "expectedAnswer",
          "difficulty",
          "expectedConcepts",
          "sourceSection"
        ]
      },
      "TechnicalNoteDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "topic": {
            "type": "string",
            "nullable": true
          },
          "rawInput": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "CONCEPT",
              "ALGORITHM",
              "SYSTEM_DESIGN",
              "BEHAVIORAL",
              "LANGUAGE_SPECIFIC"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "REVIEWING",
              "NEEDS_PRACTICE",
              "INTERVIEW_READY",
              "MASTERED",
              "PUBLISHED",
              "ARCHIVED"
            ]
          },
          "overrideRole": {
            "type": "string",
            "nullable": true
          },
          "overrideLevel": {
            "type": "string",
            "enum": [
              "JUNIOR",
              "MID",
              "SENIOR",
              "STAFF",
              "PRINCIPAL"
            ],
            "nullable": true
          },
          "overrideStack": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "overrideGoals": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "overrideEnglishLevel": {
            "type": "string",
            "enum": [
              "BEGINNER",
              "ELEMENTARY",
              "INTERMEDIATE",
              "UPPER_INTERMEDIATE",
              "ADVANCED",
              "NATIVE"
            ],
            "nullable": true
          },
          "preferredOutputStyle": {
            "type": "string",
            "nullable": true
          },
          "structuredContent": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/TechnicalNoteContentDto"
              }
            ]
          },
          "sections": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TechnicalNoteContentSectionDto"
            }
          },
          "questions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NoteGeneratedQuestionDto"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "userId",
          "title",
          "topic",
          "rawInput",
          "type",
          "status",
          "overrideRole",
          "overrideStack",
          "overrideGoals",
          "preferredOutputStyle",
          "sections",
          "questions",
          "createdAt",
          "updatedAt"
        ]
      },
      "NoteAdvancedSettingsDto": {
        "type": "object",
        "properties": {
          "targetRole": {
            "type": "string"
          },
          "targetLevel": {
            "type": "string",
            "enum": [
              "JUNIOR",
              "MID",
              "SENIOR",
              "STAFF",
              "PRINCIPAL"
            ]
          },
          "englishLevel": {
            "type": "string",
            "enum": [
              "BEGINNER",
              "ELEMENTARY",
              "INTERMEDIATE",
              "UPPER_INTERMEDIATE",
              "ADVANCED",
              "NATIVE"
            ]
          },
          "techStack": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "interviewGoals": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "preferredOutputStyle": {
            "type": "string"
          }
        }
      },
      "CreateNoteDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200
          },
          "topic": {
            "type": "string",
            "maxLength": 120,
            "nullable": true
          },
          "roughNotes": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "CONCEPT",
              "ALGORITHM",
              "SYSTEM_DESIGN",
              "BEHAVIORAL",
              "LANGUAGE_SPECIFIC"
            ],
            "default": "CONCEPT"
          },
          "advancedSettings": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/NoteAdvancedSettingsDto"
              }
            ]
          }
        },
        "required": [
          "title",
          "roughNotes"
        ]
      },
      "UpdateNoteDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200
          },
          "topic": {
            "type": "string",
            "maxLength": 120,
            "nullable": true
          },
          "roughNotes": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "CONCEPT",
              "ALGORITHM",
              "SYSTEM_DESIGN",
              "BEHAVIORAL",
              "LANGUAGE_SPECIFIC"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "REVIEWING",
              "NEEDS_PRACTICE",
              "INTERVIEW_READY",
              "MASTERED",
              "PUBLISHED",
              "ARCHIVED"
            ]
          },
          "advancedSettings": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/NoteAdvancedSettingsDto"
              }
            ]
          }
        }
      },
      "GenerateQuestionsDto": {
        "type": "object",
        "properties": {
          "count": {
            "type": "number",
            "minimum": 1,
            "maximum": 10,
            "default": 5
          },
          "difficulty": {
            "type": "string",
            "enum": [
              "EASY",
              "MEDIUM",
              "HARD",
              "EXPERT"
            ]
          }
        }
      },
...
      "ReviewQueueEntryDto": {
        "type": "object",
        "properties": {
          "item": {
            "$ref": "#/components/schemas/ReviewItemResponseDto"
          },
          "priorityScore": {
            "type": "number"
          },
          "reasons": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "item",
          "priorityScore",
          "reasons"
        ]
      },
      "ReviewQueueResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReviewQueueEntryDto"
            }
          },
          "dueCount": {
            "type": "number"
          }
        },
        "required": [
          "items",
          "dueCount"
        ]
      },
      "ReviewAttemptDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "reviewItemId": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "rating": {
            "type": "string"
          },
          "masteryBefore": {
            "type": "number"
          },
          "masteryAfter": {
            "type": "number"
          },
          "reviewIntervalDays": {
            "type": "number"
          },
          "nextReviewAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "reviewItemId",
          "userId",
          "rating",
          "masteryBefore",
          "masteryAfter",
          "reviewIntervalDays",
          "nextReviewAt",
          "createdAt"
        ]
      },
      "ReviewRatingDto": {
        "type": "object",
        "properties": {
          "rating": {
            "type": "string",
            "enum": [
              "AGAIN",
              "HARD",
              "GOOD",
              "EASY"
            ]
          }
        },
        "required": [
          "rating"
        ]
      },
      "LearningPathItemResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          },
          "actionPath": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "priorityScore": {
            "type": "number"
          },
          "availableAt": {
            "type": "string",
            "format": "date-time"
          },
          "snoozedUntil": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "startedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "completedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "skippedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "sourceReviewItemId": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "userId",
          "type",
          "title",
          "reason",
          "actionPath",
          "status",
          "priorityScore",
          "availableAt",
          "snoozedUntil",
          "startedAt",
          "completedAt",
          "skippedAt",
          "sourceReviewItemId",
          "createdAt",
          "updatedAt"
        ]
      },
      "LearningPathActionDto": {
        "type": "object",
        "properties": {
          "action": {
            "type": "string",
            "enum": [
              "start",
              "complete",
              "snooze",
              "skip"
            ]
          }
        },
        "required": [
          "action"
        ]
      },
      "WeakConceptResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "concept": {
            "type": "string"
          },
          "occurrenceCount": {
            "type": "number"
          },
          "status": {
            "type": "string"
          },
          "lastSeenAt": {
            "type": "string",
            "format": "date-time"
          },
          "sourceAnswerIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "id",
          "userId",
          "concept",
          "occurrenceCount",
          "status",
          "lastSeenAt",
          "sourceAnswerIds"
        ]
      },
      "WeakConceptStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "IMPROVING",
              "RESOLVED",
              "IGNORED"
            ]
          }
        },
        "required": [
          "status"
        ]
      },
...
    }
  }
}